+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
 ## [Unreleased Changes]
+- [Report diagnostics in `Type::from_ast`](https://github.com/amp-lang/ampc/issues/5) (report invalid types)
+
 ### [`snapshot-02-06-2023-a`](https://github.com/amp-lang/ampc/releases/tag/snapshot-02-06-2023-a)
 (initial snapshot)